requirements.txt
fastapi
uvicorn
mistralai
pymongo[srv]
python-dotenv
pydantic>=2.8.2